chore: typing animation for chat list
